+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+**[0.30.5] – 2025-07-04**  
+
+- Die Optimierung der Prognose mit gespeicherten Produktionsdaten kann nun für jede `ForecastCalcMethod` eingestellt werden.  
+  **ACHTUNG:** `ForecastCalcMethod` `median_opt` nicht mehr gültig muss nun `median+` heißen.  
+
+- Im `ForecastMgr` können nun ältere Prognosedaten gelöschti werden, falls sie nicht mehr repräsentativ sind. 
+  Einträge die älter als 35 Tage sind, werden programmtechnisch gelöscht.
+
 **[0.30.4] – 2025-06-29**  
 
 - Für den Wetterdienst von `Akkudoktor` können nun auch mehr als 2 Strings konfiguriert werden.